Refactor SaaS context import and provider placement
